Draft: move Trimex GuiCommand to gui_trimex module
